python_backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
pymysql==1.1.0
cryptography==41.0.7

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8

# Validation and serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# CORS and middleware (FastAPI has built-in CORS support)
# No additional CORS package needed

# File handling and utilities
python-magic==0.4.27
pillow==10.1.0
aiofiles==23.2.1

# HTTP requests
requests==2.31.0

# Email
fastapi-mail==1.4.1

# WebSocket support
websockets==12.0

# Background tasks
celery==5.3.4
redis==5.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
flake8==6.1.0
